ai_image_platform_v2/backend/requirements.txt
# Web框架
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
Flask-Migrate>=4.0.0

# 数据库
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
rembg>=2.0.50
onnxruntime>=1.15.0

# 认证和安全
PyJWT>=2.8.0
bcrypt>=4.0.0
Werkzeug>=2.3.0

# 缓存和会话
redis>=5.0.0
Flask-Session>=0.5.0

# HTTP客户端
requests>=2.31.0

# AI模型集成
# 通义千问API客户端
dashscope>=1.14.0
# 阿里云SDK（可选，用于更高级的功能）
alibabacloud_tea_openapi>=0.3.7
# JSON处理
orjson>=3.9.0

# 文件处理
click>=8.1.0

# 系统监控
psutil>=5.9.0

# 开发和测试
pytest>=7.4.0
pytest-flask>=1.2.0
flake8>=6.1.0
black>=23.9.0

# 生产部署
gunicorn>=21.2.0

# 工具库
python-dotenv>=1.0.0
marshmallow>=3.20.0